convert cards to ts, add run scripts
