+++ exercises/typescript/.eslintrc
@@ -10,10 +10,12 @@
     "plugin:@typescript-eslint/recommended"
   ],
   "rules": {
-    "indent": ["error", 2],
+    "indent": ["error", 2, { "SwitchCase": 1 }],
     "no-trailing-spaces": "error",
     "no-multiple-empty-lines": ["error", {"max": 1, "maxEOF": 0}],
+    "no-multi-spaces": "error",
     "object-curly-spacing": [2, "always"],
-    "semi": ["error", "never"]
+    "semi": ["error", "never"],
+    "@typescript-eslint/no-explicit-any": "error"
   }
 }
