+++ exercises/typescript/package.json
@@ -2,7 +2,7 @@
   "name": "js-fundamentals-1",
   "version": "1.0.0",
   "description": "JavaScript equivalent of the apprentice boot camp fundamentals 1",
-  "main": "src/cards/cards.js",
+  "main": "src/cards/cards.ts",
   "engines": {
     "node": ">=11.14"
   },
@@ -10,6 +10,8 @@
     "test": "test"
   },
   "scripts": {
+    "cards": "ts-node src/cards/cards.ts",
+    "snap": "ts-node src/cards/snap.ts",
     "lint": "eslint . --ext .ts",
     "test": "jest test --notify",
     "test:watch": "jest --watch"
@@ -26,6 +28,7 @@
     "jest": "^29.3.1",
     "node-notifier": "^10.0.0",
     "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
   }
 }
