+++ exercises/typescript/src/cards/animal-card.ts
@@ -1,15 +1,15 @@
 export class AnimalCard {
-  private animal: string
+  private readonly animal: string
 
   constructor(animal: string) {
     this.animal = animal
   }
 
-  snap(otherCard: AnimalCard) {
+  snap(otherCard: AnimalCard): boolean {
     return otherCard && this.animal === otherCard.animal
   }
 
-  toString() {
+  toString(): string {
     return this.animal.toString()
   }
 }
